requirements.txt
discord.py>=2.4,<3
asyncpg>=0.30,<1
python-dotenv>=1.0,<2
aiohttp>=3.11,<4